+++ lib/flutter_html.dart
@@ -17,6 +17,11 @@ class Html extends StatelessWidget {
     this.customTextStyle,
     this.blockSpacing = 14.0,
     this.useRichText = false,
+    this.onImageError,
+    this.linkStyle = const TextStyle(
+        decoration: TextDecoration.underline,
+        color: Colors.blueAccent,
+        decorationColor: Colors.blueAccent),
   }) : super(key: key);
 
   final String data;
@@ -27,6 +32,8 @@ class Html extends StatelessWidget {
   final bool renderNewlines;
   final double blockSpacing;
   final bool useRichText;
+  final ImageErrorListener onImageError;
+  final TextStyle linkStyle;
 
   /// Either return a custom widget for specific node types or return null to
   /// fallback to the default rendering.
@@ -52,6 +59,8 @@ class Html extends StatelessWidget {
                 customEdgeInsets: customEdgeInsets,
                 customTextStyle: customTextStyle,
                 html: data,
+                onImageError: onImageError,
+                linkStyle: linkStyle,
               )
             : HtmlOldParser(
                 width: width,
@@ -60,6 +69,8 @@ class Html extends StatelessWidget {
                 customRender: customRender,
                 html: data,
                 blockSpacing: blockSpacing,
+                onImageError: onImageError,
+                linkStyle: linkStyle,
               ),
       ),
     );