+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.9.7] - March 26, 2019:
+
+* Added onImageError callback
+
+## [0.9.6] - March 11, 2019:
+
+* Fix whitespace issue. ([#59](https://github.com/Sub6Resources/flutter_html/issues/59))
+
+## [0.9.5] - March 11, 2019:
+
+* Add support for `span` in `RichText` parser. ([#61](https://github.com/Sub6Resources/flutter_html/issues/61))
+* Adds `linkStyle` attribute. ([#70](https://github.com/Sub6Resources/flutter_html/pull/70))
+* Adds tests for `header`, `hr`, and `i` ([#62](https://github.com/Sub6Resources/flutter_html/issues/62))
+
 ## [0.9.4] - February 5, 2019:
 
 * Fixes `table` error in `RichText` parser. ([#58](https://github.com/Sub6Resources/flutter_html/issues/58))